+++ src/libcollections/enum_set.rs
@@ -155,7 +155,7 @@ mod test {
     use std::prelude::*;
     use std::mem;
 
-    use enum_set::{EnumSet, CLike};
+    use super::{EnumSet, CLike};
 
     #[deriving(PartialEq, Show)]
     #[repr(uint)]
