+++ src/libstd/collections/hash/map.rs
