+++ src/libstd/collections/hash/set.rs
@@ -22,7 +22,7 @@ use iter;
 use option::{Some, None};
 use result::{Ok, Err};
 
-use super::{HashMap, Entries, MoveEntries, INITIAL_CAPACITY};
+use super::map::{HashMap, Entries, MoveEntries, INITIAL_CAPACITY};
 
 
 // Future Optimization (FIXME!)
