+++ src/libstd/collections/hash/table.rs
